feat: add `setup` command
