+++ src/index.js
@@ -1,59 +1,7 @@
-#!/usr/bin/env node
+import { shouldSetup } from './args.js'
 
-import { config } from './config/index.js'
+const action = shouldSetup ? 'setup' : 'airdrop'
 
-import { banner } from './tui/banner.js'
-
-import { confirm } from './tui/confirm/index.js'
-import { Progress } from './tui/progress/index.js'
-import { Table, isEmpty, isFullfield } from './tui/table.js'
-
-import { validateAddresses } from './validator/index.js'
-
-import { checkBalance } from './lib/balance.js'
-import { airdrop } from './lib/airdrop.js'
-
-import { validateOnlyFlag } from './args.js'
+const { default: run } = await import(`./commands/${action}.js`)
 
 run()
-
-async function run() {
-  banner()
-
-  const progress = new Progress()
-
-  const addresses = await validateAddresses(progress)
-
-  const totalAddresses = addresses.length
-
-  await checkBalance(progress, totalAddresses)
-
-  if (!validateOnlyFlag) {
-    const confirmed = await confirm(
-      `Would you like to start Airdrop from ${config.address}?`
-    )
-
-    if (confirmed) {
-      const { amount, failedCount, successCount } = await airdrop(
-        progress,
-        addresses
-      )
-
-      printSummary(amount, totalAddresses, failedCount, successCount)
-    }
-  }
-}
-
-function printSummary(amount, totalAddresses, failedCount, successCount) {
-  new Table('Airdrop Summary')
-    .row('Sent a total of', `${amount} ADM`, {
-      color: isFullfield(amount, totalAddresses * config.amount)
-    })
-    .row('Failed to send to', `${failedCount} addresses`, {
-      color: isEmpty(failedCount, totalAddresses)
-    })
-    .row('Successfully sent to', `${successCount} addresses`, {
-      color: isFullfield(successCount, totalAddresses)
-    })
-    .print()
-}
