+++ src/lib/api.js
@@ -1,5 +1,5 @@
 import { AdamantApi } from 'adamant-api'
-import { config } from '../config/index.js'
+import { config } from '../utils/config/index.js'
 
 const api = new AdamantApi({
   nodes: config.nodes,
