+++ src/utils/dir.js
@@ -1,0 +1,7 @@
+import fs from 'fs'
+
+export function createDirectory(path) {
+  if (!fs.existsSync(path)) {
+    fs.mkdirSync(path, { recursive: true })
+  }
+}
