+++ src/utils/logger/output/utils.js
@@ -1,7 +1,7 @@
-import fs from 'fs'
-
 import { config } from '../../config/index.js'
-import { fullDate } from '../../utils/date.js'
+
+import { createDirectory } from '../../dir.js'
+import { fullDate } from '../../date.js'
 
 export function createOutputDirectories() {
   const outputPath = `./${config.outputPath}/${fullDate()}`
@@ -9,10 +9,4 @@ export function createOutputDirectories() {
   createDirectory(outputPath)
 
   return outputPath
-}
-
-function createDirectory(path) {
-  if (!fs.existsSync(path)) {
-    fs.mkdirSync(path, { recursive: true })
-  }
 }
